refactor: split login page into auth components, remove dead code
- Extract LoginForm/MagicLinkModal/ForgotPasswordModal to src/lib/components/auth/
- login/+page.svelte: 701 → 308 lines (Google/Passkey/guest + modal control only)
- Mode switching preserved via bind:mode (same behavior as before)
- Remove dead code: PostCard.svelte, SearchBar.svelte, check_db.ts, stores/search.ts
- Verified: svelte-check 0, eslint 0, tests 61/61, login page renders correctly
