--- src/lib/components/SearchBar.svelte
@@ -1,61 +1,0 @@
-<script lang="ts">
-	import { createEventDispatcher } from 'svelte';
-
-	const dispatch = createEventDispatcher<{ search: string }>();
-
-	export let searchQuery: string = '';
-
-	function handleSearch() {
-		dispatch('search', searchQuery);
-	}
-
-	function handleKeydown(event: KeyboardEvent) {
-		if (event.key === 'Enter') {
-			handleSearch();
-		}
-	}
-</script>
-
-<div class="relative">
-	<div class="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
-		<svg class="h-5 w-5 text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-			<path
-				stroke-linecap="round"
-				stroke-linejoin="round"
-				stroke-width="2"
-				d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-			/>
-		</svg>
-	</div>
-	<input
-		type="text"
-		bind:value={searchQuery}
-		on:keydown={handleKeydown}
-		placeholder="メモを検索..."
-		class="bg-base-100 border-base-300 focus:border-primary focus:ring-primary block w-full rounded-lg border py-2 pr-3 pl-10 leading-5 opacity-70 focus:opacity-100 focus:ring-1 focus:outline-none"
-	/>
-	{#if searchQuery}
-		<button
-			on:click={() => {
-				searchQuery = '';
-				handleSearch();
-			}}
-			class="absolute inset-y-0 right-0 flex items-center pr-3"
-			aria-label="クリア"
-		>
-			<svg
-				class="h-5 w-5 text-gray-400 hover:text-gray-600"
-				fill="none"
-				viewBox="0 0 24 24"
-				stroke="currentColor"
-			>
-				<path
-					stroke-linecap="round"
-					stroke-linejoin="round"
-					stroke-width="2"
-					d="M6 18L18 6M6 6l12 12"
-				/>
-			</svg>
-		</button>
-	{/if}
-</div>
